Move cache into core.
