+++ hawkbit-repository/hawkbit-repository-jpa/pom.xml
@@ -56,10 +56,6 @@
          <artifactId>guava</artifactId>
       </dependency>
       <dependency>
-         <groupId>org.springframework</groupId>
-         <artifactId>spring-context-support</artifactId>
-      </dependency>
-      <dependency>
          <groupId>org.springframework.boot</groupId>
          <artifactId>spring-boot-starter-data-jpa</artifactId>
       </dependency>
