--- hawkbit-repository/hawkbit-repository-jpa/src/test/java/org/eclipse/hawkbit/repository/jpa/JpaTestConfiguration.java
@@ -1,30 +1,0 @@
-/**
- * Copyright (c) 2015 Bosch Software Innovations GmbH and others.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-package org.eclipse.hawkbit.repository.jpa;
-
-import org.eclipse.hawkbit.tenancy.TenantAware;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * Overrides configurations during test (e.g.
- * {@link RepositoryApplicationConfiguration}).
- *
- */
-@Configuration
-public class JpaTestConfiguration {
-
-    /**
-     * Disables caching during test to avoid concurrency failures during test.
-     */
-    @Bean
-    RolloutStatusCache rolloutStatusCache(final TenantAware tenantAware) {
-        return new RolloutStatusCache(tenantAware, 0);
-    }
-}
